Created contract lib and moved related content in it
