+++ apps/catalog/catalog/src/app/marketplace/movie/search/filter.util.ts
@@ -1,5 +1,5 @@
 import { CatalogSearch } from './search.form';
-import { Movie } from '@blockframes/movie/+state';
+import { Movie } from '@blockframes/movie/+state/movie.model';
 import { AFM_DISABLE } from '@env';
 import { DistributionDeal } from '@blockframes/movie/distribution-deals/+state/distribution-deal.model';
