+++ apps/catalog/marketplace/marketplace/src/app/distribution-deal/+state/index.ts
@@ -2,4 +2,3 @@ export * from  './cart.model';
 export * from './cart.query';
 export * from './cart.store';
 export * from './cart.service';
-export * from './cart.firestore';
