+++ apps/catalog/marketplace/marketplace/src/app/movie/search/filter.util.ts
@@ -1,6 +1,7 @@
 import { CatalogSearch } from './search.form';
-import { Movie, DistributionDeal } from '@blockframes/movie/+state';
+import { Movie } from '@blockframes/movie/+state';
 import { AFM_DISABLE } from '@env';
+import { DistributionDeal } from '@blockframes/movie/distribution-deals/+state/distribution-deal.model';
 
 function productionYearBetween(movie: Movie, range: { from: number; to: number }): boolean {
   if (!range || !(range.from && range.to)) {
