+++ libs/contract/jest.config.js
@@ -1,0 +1,5 @@
+module.exports = {
+  name: 'contract',
+  preset: '../../jest.config.js',
+  coverageDirectory: '../../coverage/libs/contract'
+};
