+++ libs/contract/package.json
@@ -1,0 +1,8 @@
+{
+  "name": "@blockframes/contract",
+  "version": "0.0.1",
+  "peerDependencies": {
+    "@angular/common": "^8.2.13",
+    "@angular/core": "^8.2.13"
+  }
+}
