+++ libs/contract/src/index.ts
@@ -1,0 +1,5 @@
+export * from './lib/+state/contract.model';
+export * from './lib/+state/contract.query';
+export * from './lib/+state/contract.service';
+export * from './lib/+state/contract.store';
+export * from './lib/+state/contract.firestore';
