+++ libs/contract/src/lib/+state/contract.service.spec.ts
@@ -1,0 +1,14 @@
+import { ContractStore } from './contract.store';
+
+describe('ContractStore', () => {
+  let store: ContractStore;
+
+  beforeEach(() => {
+    store = new ContractStore();
+  });
+
+  it('should create an instance', () => {
+    expect(store).toBeTruthy();
+  });
+
+});
