+++ libs/contract/src/test-setup.ts
@@ -1,0 +1,2 @@
+import 'jest-preset-angular';
+import 'jest';
