+++ libs/contract/tslint.json
@@ -1,0 +1,17 @@
+{
+  "extends": "../../tslint.json",
+  "rules": {
+    "directive-selector": [
+      true,
+      "attribute",
+      "contract",
+      "camelCase"
+    ],
+    "component-selector": [
+      true,
+      "element",
+      "contract",
+      "kebab-case"
+    ]
+  }
+}
