+++ libs/material/src/lib/delivery/stakeholder/+state/stakeholder.store.ts
@@ -4,17 +4,9 @@ import { StakeholderDocument } from './stakeholder.firestore';
 
 export interface StakeholderState extends EntityState<StakeholderDocument>, MultiActiveState {}
 
-const initialState = {
-  active: []
-}
-
 @Injectable({ providedIn: 'root' })
 @StoreConfig({ name: 'stakeholder' })
 export class StakeholderStore extends EntityStore<StakeholderState, StakeholderDocument> {
-
-  constructor() {
-    super(initialState);
-  }
 
 }
