+++ libs/movie/src/lib/distribution-deals/create/availabilities.util.ts
@@ -1,7 +1,7 @@
-import { DateRange } from '@blockframes/utils/common-interfaces/date-range';
 import { MovieSalesAgentDeal } from '@blockframes/movie/+state';
 import { MovieLanguageSpecification } from '@blockframes/movie/+state/movie.firestore';
 import { DistributionDeal } from '../+state/distribution-deal.model';
+import { DateRange } from '@blockframes/utils/common-interfaces/range';
 
 /**
  * These function should be used in connection. For instance, we look for movie distribution deals in
